requirements.txt
tornado==6.4
pymysql==1.1.0
openpyxl==3.1.2
reportlab==4.0.9
python-dotenv==1.0.0
redis==5.0.1
tzdata>=2024.1
